+++ e2e/test-env/src/client/editor.jsx
@@ -3,7 +3,11 @@ import { hydrate } from 'react-dom';
 import './app.css';
 import Editor from '../shared/components/Editor';
 
-const props = { initialState: window.__CONTENT_STATE__, isMobile: window.isMobile };
+const props = {
+  initialState: window.__CONTENT_STATE__,
+  isMobile: window.isMobile,
+  locale: window.locale,
+};
 const app = <Editor {...props} />;
 
 hydrate(app, document.getElementById('root'));