+++ e2e/tests/editor.e2e.js
@@ -37,4 +37,60 @@ describe('editor', () => {
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.UNORDERED_LIST, [550, 1])
       .matchSnapshots();
   });
+
+  context('when in hebrew locale', () => {
+    beforeEach(() => cy.switchToHebrew());
+
+    context('desktop', () => {
+      beforeEach(() => cy.switchToDesktop());
+
+      it('should render plugin toolbar in rtl', () => {
+        cy.loadEditor()
+          .focusEditor()
+          .openPluginToolbar();
+      });
+
+      it('should render text toolbar in rtl', () => {
+        cy.loadEditor('plain')
+          .setSelection(0, 8)
+          .get('[data-hook=inlineToolbar]');
+      });
+
+      it('should render rtl and ltr text correctly', () => {
+        cy.loadEditor('hebrew');
+      });
+
+      it('should render external modal in rtl', () => {
+        cy.loadEditor('images')
+          .openImageSettings()
+          .get('[data-hook="imageSettingsCaptionInput"]')
+          .blur();
+      });
+
+      afterEach(() => cy.matchSnapshots({ capture: 'viewport' }));
+    });
+
+    context('mobile', () => {
+      beforeEach(() => cy.switchToMobile());
+
+      it('should render add plugin modal in rtl', () => {
+        cy.loadEditor()
+          .focusEditor()
+          .openAddPluginModal();
+      });
+
+      it('should render rtl and ltr text correctly', () => {
+        cy.loadEditor('hebrew');
+      });
+
+      it('should render external modal in rtl', () => {
+        cy.loadEditor('images')
+          .openImageSettings()
+          .get('[aria-label="Cancel"]')
+          .blur();
+      });
+
+      afterEach(() => cy.matchSnapshots({ capture: 'viewport' }));
+    });
+  });
 });