+++ examples/main/src/App.jsx
@@ -31,7 +31,7 @@ class App extends PureComponent {
     if (locale !== 'en') {
       this.setLocale(locale);
     }
-    window.onbeforeunload = function() { return "Your work will be lost."; };
+    disableBrowserBackButton();
   }
 
   getInitialState() {
@@ -246,6 +246,34 @@ class App extends PureComponent {
       </div>
     );
   }
+}
+
+function disableBrowserBackButton() {
+  (function(global) {
+    if (typeof global === 'undefined') {
+      throw new Error('window is undefined');
+    }
+
+    var _hash = '!';
+    var noBackPlease = function() {
+      global.location.href += '#';
+
+      // making sure we have the fruit available for juice (^__^)
+      global.setTimeout(function() {
+        global.location.href += '!';
+      }, 50);
+    };
+
+    global.onhashchange = function() {
+      if (global.location.hash !== _hash) {
+        global.location.hash = _hash;
+      }
+    };
+
+    global.onload = function() {
+      noBackPlease();
+    };
+  })(window);
 }
 
 export default hot(App);