+++ examples/main/src/editor/Editor.jsx
@@ -167,13 +167,11 @@ export default class Editor extends PureComponent {
           role="dialog"
           onRequestClose={onRequestClose || this.helpers.closeModal}
         >
-          {this.state.showModal && (
-            <RichContentEditorModal
-              modalsMap={ModalsMap}
-              locale={this.props.locale}
-              {...this.state.modalProps}
-            />
-          )}
+          <RichContentEditorModal
+            modalsMap={ModalsMap}
+            locale={this.props.locale}
+            {...this.state.modalProps}
+          />
         </ReactModal>
       </div>
     );