+++ packages/plugin-image/web/src/image-component.jsx
@@ -116,6 +116,7 @@ class ImageComponent extends React.Component {
     // The broadcast is good if the toolbar is displaying some status or image
     const imageData = data.length ? data[0] : data;
     this.props.componentData.src = imageData;
+    this.props.componentData.config.alignment = imageData.width >= 740 ? 'center' : 'left';
     const { setData } = this.props.blockProps;
     setData(this.props.componentData);