+++ packages/viewer/web/src/utils/convertContentState.jsx
@@ -4,7 +4,7 @@ import { convertFromRaw } from '@wix/draft-js';
 import { BLOCK_TYPES } from 'wix-rich-content-common';
 import redraft from 'redraft';
 import classNames from 'classnames';
-import { endsWith } from 'lodash';
+import { endsWith, isEmpty } from 'lodash';
 import List from '../List';
 import getPluginViewers from '../getPluginViewers';
 import { getTextDirection, kebabToCamelObjectKeys } from './textUtils';
@@ -97,6 +97,10 @@ const normalizeContentState = contentState => ({
     let text = block.text;
     if (endsWith(text, '\n')) {
       text += '\n';
+    }
+
+    if (block.type === 'unstyled' && isEmpty(text.trim())) {
+      text = '\u00A0'; // non-breaking space
     }
 
     return {